+++ CHANGELOG.md
@@ -114,6 +114,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Breaking**: `xdot_color_stop.frac` is now a `double`.
 - The GD plugin’s PNG output (`-Tpng:cairo:gd`) now sets DPI in the written PNG.
   #1617
+- The GDI+ plugin’s output (`-T*:cairo:gdiplus`) now sets DPI in the written
+  output. #1617
 
 ### Fixed